removed active flag, uses stop_time instead

+++ app/assets/javascripts/angular/controllers/lines.js
@@ -66,7 +66,7 @@ LinesCtrl = ['$scope', '$rootScope', '$http', '$location', function($scope, $roo
   $scope.activePeople = function(line) {
     var cnt = 0;
     $.each(line.people, function(i, person) {
-      if(person.active)
+      if(!person.stop_time)
         cnt++;
     });
     return cnt;
@@ -76,7 +76,7 @@ LinesCtrl = ['$scope', '$rootScope', '$http', '$location', function($scope, $roo
     var p;
     if(line.people) {
       $.each(line.people, function(i, person) {
-        if(person.active) {
+        if(!person.stop_time) {
           p = person;
           return false;
         }
@@ -103,8 +103,7 @@ LinesCtrl = ['$scope', '$rootScope', '$http', '$location', function($scope, $roo
   $scope.enqueue = function($event, line) {
     $event.stopImmediatePropagation();
     var queueObj = {
-      line_id: line.id,
-      active: true
+      line_id: line.id
     }
     $http.post('/person', queueObj).success(function(data) {
       queueObj.id = parseInt(data.id);
@@ -118,15 +117,17 @@ LinesCtrl = ['$scope', '$rootScope', '$http', '$location', function($scope, $roo
     $event.stopImmediatePropagation();
     var id;
     $.each(line.people, function(i, person) {
-      if(person.active) {
+      if(!person.stop_time) {
         id = person.id;
-        person.active = false;
+        person.stop_time = new Date();
         return false;
       }
     });
-    $http.delete('/person/' + id).success(function(data) {
-      console.log('dequeued');
-    });
+    if(id) {
+      $http.delete('/person/' + id).success(function(data) {
+        console.log('dequeued');
+      });
+    }
   }
 
   $scope.isPresent = function(line) {
